Firebase Authentication Mashup
Trying to get firebase authorization integrated. Here is what I have currently on my local.

The broken_index.html file is the file with the issue that refreshes uncontrollably and the index.html file is the file with the UI that works but has too much going on in the middle. I'd prefer to get the broken_index.html file up working.